Cleanup before handing in project.
Fixed the save method to check if the name is empty.
Fixed the messageboxes to show if the value is less than 1 with a added while loop.

Fixed better names for save and loading in falehandler. Increased readability.

Added a filter if the points x1, x2 are equal and y1, y2 are equal then the shape should not be added.